FS-674 XDI OM config service REST access changes
This is a fairly large set of changes to finish up the work to enable cluster configuration and events for the XDI.  The bulk of this commit is in 2 main parts (that are intertwined enough that I couldn't easily do 2 separate commits):

1. Xdi CachedConfiguration changes to update the cache directly as changes are made rather than clearing and reloading the cache any time a change is made.  The cache updater still runs, but on a slightly longer interval, in order to update when it detects changes on the configdb (configuration version updated).  This is necessary to capture changes made directly through the OM that do not go through XDI, such as UI access.

2. Secondly, this change adds a lot of trace-level logging that was useful for debugging authentication/authorization issues that were occurring when running the tests.  In the past we did not log anything on auth(z) failures to avoid exposing security information.  However, this made it very difficult to track down errors that turned out to be due to a problem introduced into the tenantId lookup that was manifesting itself as an authorization failure.  Exceptions that ultimately go to end users still remain free of potentially identifying security information.  We currently default the log level to trace, so these messages will be logged.  In production the log level should be turned way down to INFO or WARN so they will not be an impact.

+++ source/java/src/com/formationds/xdi/Xdi.java
@@ -42,7 +42,7 @@ public class Xdi {
     }
 
     public long createVolume(AuthenticationToken token, String domainName, String volumeName, VolumeSettings volumePolicy) throws ApiException, TException {
-        config.createVolume(domainName, volumeName, volumePolicy, authorizer.tenantId(token));
+        config.createVolume( domainName, volumeName, volumePolicy, authorizer.tenantId( token ) );
 
         /**
          * allows the UI to assign a snapshot policy to a volume without having to make an
@@ -51,9 +51,13 @@ public class Xdi {
         return config.getVolumeId( volumeName );
     }
 
+    // TODO: this is ignoring case where volume doesn't exist.  s3 apis may require the return of a
+    // specific error code indicating the bucket does not exist.
     public void deleteVolume(AuthenticationToken token, String domainName, String volumeName) throws ApiException, TException {
-        attemptVolumeAccess(token, volumeName);
-        config.deleteVolume(domainName, volumeName);
+        if ( config.statVolume( domainName, volumeName ) != null ) {
+            attemptVolumeAccess( token, volumeName );
+            config.deleteVolume( domainName, volumeName );
+        }
     }
 
     public VolumeStatus statVolume(AuthenticationToken token, String domainName, String volumeName) throws ApiException, TException {
